refactor(cketh): give the pipeline's request contract its own file
`transactions/mod.rs` was 1562 lines covering six concerns. Move the smallest
self-contained one out: `PipelineRequest` — what a request must answer to travel a
`TransactionPipeline` — together with the minter's implementation of it for
`WithdrawalRequest`, which is most of the bulk (the two `create_transaction` arms).

`mod.rs` drops to 1384 lines and loses three imports that existed only for the
transaction-building code (`EthereumNetwork`, `GasAmount`, `ResubmissionStrategy`).
The trait is re-exported, so no caller outside the module changes.

One test was reaching `EthereumNetwork` through `state::transactions`, which worked
only because a child module can see its parent's private `use`. It now imports from
`crate::lifecycle`, where the type actually lives.
